+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Temporary user creation during dogu start (#88)
+- Permissions of the temporary admin user used to import quality profiles (#88)
+
+### Changed
+- Blocked updates from versions prior to 8.x
 
 ## [v9.9.1-3] - 2023-05-15
 ### Changed